where clause

+++ sql/src/main/scala/unicorn/sql/SQLAST.scala
@@ -104,7 +104,7 @@ trait LiteralExpression extends Expression {
   override def isLiteral = true
   def gatherFields = Seq.empty
 }
-case class IntLiteral(v: Long) extends LiteralExpression
+case class IntLiteral(v: Int) extends LiteralExpression
 case class FloatLiteral(v: Double) extends LiteralExpression
 case class StringLiteral(v: String) extends LiteralExpression
 case class NullLiteral() extends LiteralExpression
